Complete Phase 3: Habits domain extraction and split
- Add CalculateActivityDataUseCase for activity calculation logic
- Add PrewarmActivityDataUseCase for parallel computation
- Add command grouping (Memo, Refresh, Activity)
- Split HabitsEffectHandler into 3 focused sub-handlers
- Update factory and tests to use new architecture
- Replace flow{} with actions{}/sideEffect{} helpers

HabitsEffectHandler reduced from 247 to 17 lines
