+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/habits/di/HabitsFeatureFactory.kt
@@ -2,9 +2,14 @@ package space.be1ski.vibits.shared.feature.habits.di
 
 import space.be1ski.vibits.shared.core.elm.Feature
 import space.be1ski.vibits.shared.core.elm.FeatureImpl
+import space.be1ski.vibits.shared.feature.habits.domain.usecase.CalculateActivityDataUseCase
+import space.be1ski.vibits.shared.feature.habits.domain.usecase.PrewarmActivityDataUseCase
 import space.be1ski.vibits.shared.feature.habits.presentation.HabitsAction
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsActivityEffectHandler
 import space.be1ski.vibits.shared.feature.habits.presentation.HabitsEffect
 import space.be1ski.vibits.shared.feature.habits.presentation.HabitsEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsMemoEffectHandler
+import space.be1ski.vibits.shared.feature.habits.presentation.HabitsRefreshEffectHandler
 import space.be1ski.vibits.shared.feature.habits.presentation.HabitsState
 import space.be1ski.vibits.shared.feature.habits.presentation.habitsReducer
 
@@ -21,9 +26,29 @@ fun createHabitsFeature(
     reducer = habitsReducer,
     effectHandler =
       HabitsEffectHandler(
-        memosRepository = dependencies.memosRepository,
-        onRefresh = onRefresh,
-        buildActivityDataUseCase = dependencies.buildActivityDataUseCase,
-        calculateSuccessRateUseCase = dependencies.calculateSuccessRateUseCase,
+        memoHandler =
+          HabitsMemoEffectHandler(
+            memosRepository = dependencies.memosRepository,
+          ),
+        refreshHandler =
+          HabitsRefreshEffectHandler(
+            onRefresh = onRefresh,
+          ),
+        activityHandler =
+          HabitsActivityEffectHandler(
+            calculateActivityDataUseCase =
+              CalculateActivityDataUseCase(
+                buildActivityDataUseCase = dependencies.buildActivityDataUseCase,
+                calculateSuccessRateUseCase = dependencies.calculateSuccessRateUseCase,
+              ),
+            prewarmActivityDataUseCase =
+              PrewarmActivityDataUseCase(
+                calculateActivityDataUseCase =
+                  CalculateActivityDataUseCase(
+                    buildActivityDataUseCase = dependencies.buildActivityDataUseCase,
+                    calculateSuccessRateUseCase = dependencies.calculateSuccessRateUseCase,
+                  ),
+              ),
+          ),
       ),
   )
