+++ shared/src/commonMain/kotlin/space/be1ski/vibits/shared/feature/habits/presentation/HabitsRefreshEffectHandler.kt
@@ -1,0 +1,23 @@
+package space.be1ski.vibits.shared.feature.habits.presentation
+
+import kotlinx.coroutines.flow.Flow
+import space.be1ski.vibits.shared.core.elm.EffectHandler
+import space.be1ski.vibits.shared.core.elm.sideEffect
+import space.be1ski.vibits.shared.core.logging.Log
+
+private const val TAG = "HabitsRefreshEffect"
+
+class HabitsRefreshEffectHandler(
+  private val onRefresh: () -> Unit,
+) : EffectHandler<HabitsEffect.Refresh, HabitsAction> {
+  override fun invoke(effect: HabitsEffect.Refresh): Flow<HabitsAction> =
+    when (effect) {
+      HabitsEffect.RefreshMemos -> handleRefreshMemos()
+    }
+
+  private fun handleRefreshMemos(): Flow<HabitsAction> =
+    sideEffect {
+      Log.d(TAG, "Refreshing memos")
+      onRefresh()
+    }
+}
